Rename answer-related files
