+++ src/server/models/index.ts
@@ -1,5 +1,5 @@
 import * as questions from "./question";
-import * as answers from "./answers";
+import * as answers from "./answer";
 import * as users from "./user";
 
 export { questions, answers, users };
