+++ src/server/routes/index.ts
@@ -1,5 +1,5 @@
 import questions from "./question";
-import answers from "./answers";
+import answers from "./answer";
 import users from "./user";
 
 export { questions, answers, users };
